Add a ranked queue implementation

+++ assign2/src/main/java/pt/up/fe/cpd/proj2/server/auth/UserInfo.java
@@ -1,6 +1,6 @@
 package pt.up.fe.cpd.proj2.server.auth;
 
-public record UserInfo(int id, String username, String password, int elo) implements Comparable<UserInfo> {
+public record UserInfo(int id, String username, String password, double elo) implements Comparable<UserInfo> {
     @Override
     public int compareTo(UserInfo o) {
         return Integer.compare(id, o.id);
